refactor: update after review
