+++ src/test/java/pages/HomePage.java
@@ -1,0 +1,40 @@
+package pages;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import org.openqa.selenium.support.PageFactory;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
+import java.time.Duration;
+
+
+public class HomePage {
+
+
+    public static WebDriver driver;
+
+    public HomePage(WebDriver driver) {
+        HomePage.driver = driver;
+        PageFactory.initElements(driver, this);
+    }
+
+
+    @FindBy(xpath = "//a")
+    private WebElement logOut;
+
+
+    public String getLogoutLinkText() {
+        return logOut.getText();
+    }
+
+    public void clickLogoutBtn() {
+        logOut.click();
+    }
+
+    public static void waitHomePageLoad(String url) {
+        WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(10));
+        wait.until(ExpectedConditions.urlToBe(url));
+    }
+
+}
