--- src/test/java/utils/WebdriverSetting.java
@@ -1,28 +1,0 @@
-package utils;
-
-import org.openqa.selenium.WebDriver;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import java.time.Duration;
-
-
-public class WebdriverSetting {
-
-    public WebDriver driver;
-
-    @BeforeEach
-    public void setUp() {
-
-        ChromeOptions options = new ChromeOptions();
-        options.addArguments("--remote-allow-origins=*");
-        System.setProperty("webdriver.chrome.driver", "C:\\webdrivers\\chromedriver.exe");
-        driver = new ChromeDriver(options);
-        WebDriver.Timeouts timeouts = driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
-    }
-
-    @AfterEach
-    public void close() { driver.quit(); }
-
-}
